requirements.txt
aiogram==3.0.0
python-dotenv
telethon==1.34.0
